Changed up some stuff

+++ xe300.sh
@@ -92,21 +92,16 @@ CURRENT_HASH=$(
         jq --raw-output '.sha'
 )
 
-if [ -f "${HASH_STORE}" ]; then
+if [ -f "${HASH_STORE}" ] && [ ! -z "${CURRENT_HASH}" ]; then
     log_say "Update Script Found ${HASH_STORE}"
     CHECK_HASH=$(cat ${HASH_STORE})
     log_say "Update Script Check Hash ${CHECK_HASH}"
     [[ "${CHECK_HASH}" != "${CURRENT_HASH}" ]] && {
         log_say "Update Script ${CHECK_HASH} != ${CURRENT_HASH}"
         UPDATE_NEEDED="1"
-        echo "${CURRENT_HASH}" > "${HASH_STORE}"
-        log_say "Update Script Wrote ${CURRENT_HASH} > ${HASH_STORE}"
     }
 else
     log_say "Update Script ${HASH_STORE} did not exist"
-    touch "${HASH_STORE}"
-    echo "${CURRENT_HASH}" > "${HASH_STORE}"
-    log_say "Update Script Wrote ${CURRENT_HASH} > ${HASH_STORE}"
     UPDATE_NEEDED="1"
 fi
 
@@ -133,26 +128,38 @@ if [[ "${UPDATE_NEEDED}" == "1" || ! -d ${UPDATER_LOCATION} ]]; then
     }
 
     log_say "Update Script Cloning ${GIT_URL} into ${TMP_DIR}"
-    git clone --depth=1 "${GIT_URL}" "${TMP_DIR}"
+    while ! git clone --depth=1 "${GIT_URL}" "${TMP_DIR}" >/dev/null 2>&1; do
+        log_say "... Waiting to clone the update script repo ..."
+        sleep 1
+    done
+    # Verify it downloaded successfully
+    if [ -d "${TMP_DIR}" ]; then    
+        log_say "Update Script Cleaning up .git folder"
+        rm -rf "${TMP_DIR}/.git"
 
-    log_say "Update Script Cleaning up .git folder"
-    rm -rf "${TMP_DIR}/.git"
+        [ -d "${UPDATER_LOCATION}" ] && { log_say "Update Script Removing old ${UPDATER_LOCATION}"; rm -rf "${UPDATER_LOCATION}"; }
 
-    [ -d "${UPDATER_LOCATION}" ] && { log_say "Update Script Removing old ${UPDATER_LOCATION}"; rm -rf "${UPDATER_LOCATION}"; }
+        log_say "Update Script Moving ${TMP_DIR} to ${UPDATER_LOCATION}"
+        mv "${TMP_DIR}" "${UPDATER_LOCATION}"
 
-    log_say "Update Script Moving ${TMP_DIR} to ${UPDATER_LOCATION}"
-    mv "${TMP_DIR}" "${UPDATER_LOCATION}"
+        echo "${CURRENT_HASH}" > "${HASH_STORE}"
+        log_say "Update Script Wrote ${CURRENT_HASH} > ${HASH_STORE}"
 
-    [ -f "${UPDATER_LOCATION}/crontabs" ] && {
-        log_say "Update Script Inserting crontabs for updaters and restarting the cron service"
-        cat "${UPDATER_LOCATION}/crontabs" >> /etc/crontabs/root
-        /etc/init.d/cron restart
-    }
+        [ -f "${UPDATER_LOCATION}/crontabs" ] && {
+            log_say "Update Script Inserting crontabs for updaters and restarting the cron service"
+            cat "${UPDATER_LOCATION}/crontabs" >> /etc/crontabs/root
+            /etc/init.d/cron restart
+        }
 
-    [ -f "${UPDATER_LOCATION}/first-run.sh" ] && {
-        log_say "Running the commands in the first-run.sh script."
-        bash "${UPDATER_LOCATION}/first-run.sh"
-    }
+        [ -f "${UPDATER_LOCATION}/first-run.sh" ] && {
+            log_say "Running the commands in the first-run.sh script."
+            bash "${UPDATER_LOCATION}/first-run.sh"
+        }
+    else
+        log_say "We were not able to download our update scripts"
+        #reboot
+        exit 1
+    fi
 else
     log_say "Update Script Update is not needed"
 fi # UPDATE_NEEDED check
@@ -169,11 +176,6 @@ wait_for_opkg() {
 # Wait for opkg to finish
 wait_for_opkg
 
-if ! ; then
-    log_say "opkg update did not run correctly so we exit!"
-    exit 1
-fi
-
 while ! opkg update >/dev/null 2>&1; do
     log_say "... Waiting for opkg update to succesfully run ..."
     sleep 1
@@ -182,7 +184,7 @@ done
 # Install system packages as needed
 log_say "Checking Required Packages..."
 
-PACKAGE_LIST="modemmanager kmod-usb-serial kmod-usb-net kmod-usb-serial-wwan kmod-usb-serial-option kmod-usb-net-qmi-wwan kmod-usb-net-cdc-mbim luci-proto-modemmanager tgrouterappstore luci-app-shortcutmenu luci-app-poweroff luci-app-wizard"  # List of packages separated by space
+PACKAGE_LIST="modemmanager kmod-usb-serial kmod-usb-net kmod-usb-serial-wwan kmod-usb-serial-option kmod-usb-net-qmi-wwan kmod-usb-net-cdc-mbim luci-proto-modemmanager luci-app-shortcutmenu luci-app-poweroff luci-app-wizard"  # List of packages separated by space
 
 for package in $PACKAGE_LIST; do
     if ! opkg list-installed | grep -q "^$package -"; then
@@ -217,7 +219,8 @@ else
         uci set network.wwan.auth='none'
         uci set network.wwan.iptype='ipv4v6'
         uci set network.wwan.device="${usb_device_path}"
-        uci add_list firewall.wan.network='wwan'
+        # End user can add the interface to the WAN firewall zone to use it for Internet access
+        #uci add_list firewall.wan.network='wwan'
         uci commit
         
         log_say "Interface 'wwan' created."
